Code built at 0153d97

+++ cli/commands/build.js
@@ -4,13 +4,17 @@ const dedent = require('dedent-js');
 const shell = require('shelljs');
 const buildProject = require('./build/build_project');
 let chalk = require('chalk');
+let { getDefaultTopicAndPath } = require('./shared/helpers');
 
 function build(options) {
   let { symlinks, projectPathPrefix, hashUrls, keepBuildDirectory, manualHtml, logging } = options;
+  let { defaultTopicFilePath, defaultTopicName } = getDefaultTopicAndPath();
   if (!fs.existsSync('./topics')) throw new Error('There must be a topics directory present, try running "canopy init"');
   if (!fs.existsSync('./.canopy_default_topic')) throw new Error('There must be a default topic dotfile present, try running "canopy init"');
+  if (!defaultTopicFilePath || !fs.existsSync(defaultTopicFilePath)) {
+    throw new Error(chalk.red(`Error: No topic file corresponds to name given in .canopy_default_topic: "${defaultTopicName.trim()}"`));
+  }
 
-  let defaultTopicString = fs.readFileSync('.canopy_default_topic').toString().trim();
   let canopyLocation = process.env.CANOPY_LOCATION || path.dirname(path.dirname(fs.realpathSync(shell.which('canopy').stdout)));
 
   if (!keepBuildDirectory) fs.rmSync('build', { recursive: true, force: true });
@@ -29,7 +33,7 @@ function build(options) {
       <body>
       <div
         id="_canopy"
-        data-default-topic="${defaultTopicString}"
+        data-default-topic="${defaultTopicName}"
         data-project-path-prefix="${projectPathPrefix||''}"
         data-hash-urls="${hashUrls || ''}">
       </div>
@@ -40,7 +44,13 @@ function build(options) {
     fs.writeFileSync('build/index.html', html);
   }
 
-  buildProject('.', defaultTopicString, options);
+  if (options.logging) console.log(chalk.cyan(
+    `Canopy build: Rebuilding JSON at ${''
+    + (new Date()).toLocaleTimeString()} (pid ${process.pid})`
+    + (options.filesEdited ? ` – file changed: ${options.filesEdited}` : '')
+  ));
+
+  buildProject('.', defaultTopicName, options);
 
   if (fs.existsSync(`assets`)) {
     fs.copySync('assets', 'build/_assets', { overwrite: true });
@@ -63,7 +73,7 @@ function build(options) {
   }
 
   if (!fs.existsSync(`${canopyLocation}/dist/canopy.js`)) {
-    throw new Error('No Canopy.js asset found');
+    throw new Error(chalk.red('No Canopy.js asset found'));
   }
 
   fs.copyFileSync(`${canopyLocation}/dist/canopy.js`, 'build/canopy.js');
@@ -72,7 +82,7 @@ function build(options) {
     fs.copyFileSync(`${canopyLocation}/dist/canopy.js.map`, 'build/canopy.js.map');
   }
 
-  if (options.logging) console.log(chalk.cyan(`Canopy watch: Rebuilt JSON at ${'' + (new Date()).toLocaleTimeString()} (pid ${process.pid})` + (options.filesEdited ? ` – file changed: ${options.filesEdited}` : '')));
+  if (options.logging) console.log(chalk.cyan(`Canopy build: build finished at ${'' + (new Date()).toLocaleTimeString()} (pid ${process.pid})`));
 }
 
 function getDirectories(path) {
